fixing local deployment failure
