+++ com.example.BatteryAwareHelloWorld/gdk-config.json
@@ -2,7 +2,7 @@
     "component": {
       "com.example.BatteryAwareHelloWorld": {
         "author": "Synapse Automota",
-        "version": "0.0.1",
+        "version": "0.0.2",
         "build": {
           "build_system" : "zip"
         },
