+++ com.example.pubsubtest/gdk-config.json
@@ -1,15 +1,15 @@
 {
   "component": {
-    "com.example.pubsubtest/": {
+    "com.example.pubsubtest": {
       "author": "Synapse Automota",
-      "version": "0.0.1",
+      "version": "0.0.4",
       "build": {
         "build_system": "custom",
         "custom_build_command": [
           "bash",
           "build-custom.sh",
-          "com.example.pubsubtest/",
-          "0.0.1"
+          "com.example.pubsubtest",
+          "0.0.4"
         ]
       },
       "publish": {
