+++ pubsubpython/.gitignore
@@ -1,0 +1,11 @@
+*build/
+build
+*dist/
+*.egg-info
+*__pycache__
+*htmlcov/
+*.coverage
+*.iml
+*.DS_Store
+*.eggs
+zip-build
