+++ pubsubpython/gdk-config.json
@@ -1,0 +1,16 @@
+{
+  "component": {
+    "com.example.PythonLocalPubSub": {
+      "author": "<PLACEHOLDER_AUTHOR>",
+      "version": "NEXT_PATCH",
+      "build": {
+        "build_system": "zip"
+      },
+      "publish": {
+        "bucket": "greengrass-component-artifacts",
+        "region": "ap-southeast-2"
+      }
+    }
+  },
+  "gdk_version": "1.0.0"
+}
