+++ pubsubpython/recipe.yaml
@@ -1,0 +1,27 @@
+---
+RecipeFormatVersion: "2020-01-25"
+ComponentName: "{COMPONENT_NAME}"
+ComponentVersion: "{COMPONENT_VERSION}"
+ComponentDescription: "This example component shows how to pubslish and subscribe to topics from a component using IPC operations."
+ComponentPublisher: "{COMPONENT_AUTHOR}"
+ComponentConfiguration:
+  DefaultConfiguration:
+    Message: "Hello from local pubsub topic"
+    Topic: "/topic/local/pubsub"
+    accessControl:
+      aws.greengrass.ipc.pubsub:
+        com.example.PythonLocalPubSub:pubsub:1:
+          policyDescription: Allows access to publish and subscribe to topic /topic/local/pubsub.
+          operations:
+            - aws.greengrass#SubscribeToTopic
+            - aws.greengrass#PublishToTopic
+          resources:
+            - "/topic/local/pubsub"
+Manifests:
+  - Platform:
+      os: all
+    Artifacts:
+      - URI: "s3://greengrass-component-artifacts/pubsubpython/ap-southeast-2/LocalPubSub.zip"
+        Unarchive: ZIP
+    Lifecycle:
+      Run: "python3 -u {artifacts:decompressedPath}/LocalPubSub/main.py {configuration:/Topic} {configuration:/Message}"
